refactor: use `copy()` from `std/io`
