+++ archive/untar_test.ts
@@ -9,7 +9,7 @@ import {
   Untar,
 } from "./untar.ts";
 import { Buffer } from "../io/buffer.ts";
-import { copy } from "../streams/copy.ts";
+import { copy } from "../io/copy.ts";
 import { readAll } from "../io/read_all.ts";
 import { filePath, testdataDir } from "./_test_common.ts";
