+++ io/limited_reader_test.ts
@@ -2,7 +2,7 @@
 import { assertEquals } from "../assert/mod.ts";
 import { LimitedReader } from "./limited_reader.ts";
 import { StringWriter } from "./string_writer.ts";
-import { copy } from "../streams/copy.ts";
+import { copy } from "./copy.ts";
 import { readAll } from "./read_all.ts";
 import { StringReader } from "./string_reader.ts";
